broken: fix known bugs.
