+++ .fleet/receipt.json
@@ -1,0 +1,14 @@
+// Project generated by Kotlin Multiplatform Wizard
+{
+    "spec": {
+        "template_id": "kmt",
+        "targets": {
+            "desktop": {
+                "ui": [
+                    "compose"
+                ]
+            }
+        }
+    },
+    "timestamp": "2025-02-14T06:21:53.818113046Z"
+}
