--- composeApp/src/desktopMain/kotlin/icu/bluedream/gameinstaller/data/Constants.kt
@@ -1,18 +1,0 @@
-package icu.bluedream.gameinstaller.data
-
-object Constants {
-    const val GENSHIN_PKG_API_CN = "https://hyp-api.mihoyo.com/hyp/hyp-connect/api/getGamePackages?launcher_id=jGHBHlcOq1&game_ids[]=1Z8W5NHUQb"
-    const val GENSHIN_PKG_API_GL = "https://sg-hyp-api.hoyoverse.com/hyp/hyp-connect/api/getGamePackages?launcher_id=VYTpXlbWo8&game_ids[]=gopR6Cufr3"
-    const val ZZZ_PKG_API_CN = "https://hyp-api.mihoyo.com/hyp/hyp-connect/api/getGamePackages?launcher_id=jGHBHlcOq1&game_ids[]=x6znKlJ0xK"
-    const val ZZZ_PKG_API_GL = "https://sg-hyp-api.hoyoverse.com/hyp/hyp-connect/api/getGamePackages?launcher_id=VYTpXlbWo8&game_ids[]=U5hbdsT9W7"
-
-    const val GS_EXE = "YuanShen.exe"
-    const val ZZZ_EXE = "ZenlessZoneZero.exe"
-
-    const val WINE_DRIVE_C_PATH = "Contents/SharedSupport/prefix/drive_c/"
-    const val LAUNCH_BAT = "launchGame.bat"
-    const val WINE_MAC_ROOT_PATH = "Z:"
-    const val MAC_APPLICATION_PATH = "/Applications/"
-
-    const val GS_CLOUD_GAME_PARAM = "-platform_type CLOUD_THIRD_PARTY_PC"
-}
