--- composeApp/src/desktopMain/kotlin/icu/bluedream/gameinstaller/data/state/HomeUiState.kt
@@ -1,9 +1,0 @@
-package icu.bluedream.gameinstaller.data.state
-
-import icu.bluedream.gameinstaller.data.types.GameType
-
-data class HomeUiState(
-    var selectedGame: GameType = GameType.GI_CN,
-    var selectedPath: String = "",
-    var isInstalled: Boolean = false
-)
