--- composeApp/src/desktopMain/kotlin/icu/bluedream/gameinstaller/data/types/ActionState.kt
@@ -1,5 +1,0 @@
-package icu.bluedream.gameinstaller.data.types
-
-enum class ActionState {
-    ONGOING, WAITING, FINISHED, FAILED
-}
