+++ composeApp/src/desktopMain/kotlin/icu/bluedream/gamesteup/core/utils/ContentUtil.kt
@@ -1,0 +1,64 @@
+package icu.bluedream.gamesteup.core.utils
+
+import icu.bluedream.gamesteup.data.*
+import icu.bluedream.gamesteup.data.types.GameType
+import java.io.File
+
+object ContentUtil {
+    fun reformatMd5File(inputPath: String, outputPath: String) {
+        val regex = Regex("""MD5 \(\./(.*)\) = (.*)""")
+        File(outputPath).printWriter().use { writer ->
+            File(inputPath).forEachLine { line ->
+                val transformed = line.replace(regex) { match ->
+                    "${match.groupValues[1]}=${match.groupValues[2]}"
+                }
+                writer.println(transformed)
+            }
+        }
+    }
+
+    fun getLeftPartsList(filePath: String): MutableList<String> {
+        return File(filePath).useLines { lines ->
+            lines.map { it.substringBefore('=').trim() }
+                .filter { it.isNotEmpty() }
+                .toMutableList()
+        }
+    }
+
+    fun readMd5FileToMap(filePath: String): Map<String, String> {
+        return try {
+            File(filePath).useLines { lines ->
+                lines
+                    .map { it.trim() }
+                    .filter { it.isNotBlank() }
+                    .mapNotNull { line ->
+                        val parts = line.split('=', limit = 2)
+                            .map { it.trim() }
+                        when {
+                            parts.size != 2 -> {
+                                null
+                            }
+                            parts[0].isEmpty() -> {
+                                null
+                            }
+                            parts[1].isEmpty() -> {
+                                null
+                            }
+                            else -> parts[0] to parts[1]
+                        }
+                    }
+                    .toMap()
+            }
+        } catch (e: Exception) {
+            emptyMap()
+        }
+    }
+
+    fun modifyLaunchBat(gameType: GameType, installPath: String) {
+        fun isGS(type: GameType) = type == GameType.GI_CN || type == GameType.GI_OS
+        val launchBatPath = "$MAC_APPLICATION_PATH${GameType.getAppTranslationName(gameType)}/$WINE_DRIVE_C_PATH$LAUNCH_BAT"
+        val launchString = "\"$WINE_MAC_ROOT_PATH$installPath${GameType.getExeAppName(gameType)}\"" + if (isGS(gameType)) " $GS_CLOUD_GAME_PARAM" else ""
+        val launchBatFile = File(launchBatPath)
+        launchBatFile.writeText(launchString)
+    }
+}
