+++ composeApp/src/desktopMain/kotlin/icu/bluedream/gamesteup/data/state/HomeUiState.kt
@@ -1,0 +1,9 @@
+package icu.bluedream.gamesteup.data.state
+
+import icu.bluedream.gamesteup.data.types.GameType
+
+data class HomeUiState(
+    var selectedGame: GameType = GameType.GI_CN,
+    var selectedPath: String = "",
+    var isInstalled: Boolean = false
+)
