+++ composeApp/src/desktopMain/kotlin/icu/bluedream/gamesteup/data/types/ActionState.kt
@@ -1,0 +1,5 @@
+package icu.bluedream.gamesteup.data.types
+
+enum class ActionState {
+    ONGOING, WAITING, FINISHED, FAILED
+}
