+++ composeApp/src/desktopMain/kotlin/icu/bluedream/gamesteup/ui/theme/Color.kt
@@ -1,0 +1,74 @@
+package icu.bluedream.gamesteup.ui.theme
+import androidx.compose.ui.graphics.Color
+
+val primaryLight = Color(0xFF566422)
+val onPrimaryLight = Color(0xFFFFFFFF)
+val primaryContainerLight = Color(0xFFD9EB99)
+val onPrimaryContainerLight = Color(0xFF3F4C0B)
+val secondaryLight = Color(0xFF5C6146)
+val onSecondaryLight = Color(0xFFFFFFFF)
+val secondaryContainerLight = Color(0xFFE0E6C3)
+val onSecondaryContainerLight = Color(0xFF444930)
+val tertiaryLight = Color(0xFF3A665E)
+val onTertiaryLight = Color(0xFFFFFFFF)
+val tertiaryContainerLight = Color(0xFFBDECE1)
+val onTertiaryContainerLight = Color(0xFF214E46)
+val errorLight = Color(0xFFBA1A1A)
+val onErrorLight = Color(0xFFFFFFFF)
+val errorContainerLight = Color(0xFFFFDAD6)
+val onErrorContainerLight = Color(0xFF93000A)
+val backgroundLight = Color(0xFFFBFAED)
+val onBackgroundLight = Color(0xFF1B1C15)
+val surfaceLight = Color(0xFFFBFAED)
+val onSurfaceLight = Color(0xFF1B1C15)
+val surfaceVariantLight = Color(0xFFE3E4D3)
+val onSurfaceVariantLight = Color(0xFF46483C)
+val outlineLight = Color(0xFF77786A)
+val outlineVariantLight = Color(0xFFC7C8B8)
+val scrimLight = Color(0xFF000000)
+val inverseSurfaceLight = Color(0xFF303129)
+val inverseOnSurfaceLight = Color(0xFFF2F1E5)
+val inversePrimaryLight = Color(0xFFBDCE80)
+val surfaceDimLight = Color(0xFFDBDBCF)
+val surfaceBrightLight = Color(0xFFFBFAED)
+val surfaceContainerLowestLight = Color(0xFFFFFFFF)
+val surfaceContainerLowLight = Color(0xFFF5F4E8)
+val surfaceContainerLight = Color(0xFFEFEEE2)
+val surfaceContainerHighLight = Color(0xFFE9E9DC)
+val surfaceContainerHighestLight = Color(0xFFE4E3D7)
+
+val primaryDark = Color(0xFFBDCE80)
+val onPrimaryDark = Color(0xFF2A3500)
+val primaryContainerDark = Color(0xFF3F4C0B)
+val onPrimaryContainerDark = Color(0xFFD9EB99)
+val secondaryDark = Color(0xFFC4CAA9)
+val onSecondaryDark = Color(0xFF2E331B)
+val secondaryContainerDark = Color(0xFF444930)
+val onSecondaryContainerDark = Color(0xFFE0E6C3)
+val tertiaryDark = Color(0xFFA1D0C5)
+val onTertiaryDark = Color(0xFF033730)
+val tertiaryContainerDark = Color(0xFF214E46)
+val onTertiaryContainerDark = Color(0xFFBDECE1)
+val errorDark = Color(0xFFFFB4AB)
+val onErrorDark = Color(0xFF690005)
+val errorContainerDark = Color(0xFF93000A)
+val onErrorContainerDark = Color(0xFFFFDAD6)
+val backgroundDark = Color(0xFF13140D)
+val onBackgroundDark = Color(0xFFE4E3D7)
+val surfaceDark = Color(0xFF13140D)
+val onSurfaceDark = Color(0xFFE4E3D7)
+val surfaceVariantDark = Color(0xFF46483C)
+val onSurfaceVariantDark = Color(0xFFC7C8B8)
+val outlineDark = Color(0xFF909283)
+val outlineVariantDark = Color(0xFF46483C)
+val scrimDark = Color(0xFF000000)
+val inverseSurfaceDark = Color(0xFFE4E3D7)
+val inverseOnSurfaceDark = Color(0xFF303129)
+val inversePrimaryDark = Color(0xFF566422)
+val surfaceDimDark = Color(0xFF13140D)
+val surfaceBrightDark = Color(0xFF393A31)
+val surfaceContainerLowestDark = Color(0xFF0D0F08)
+val surfaceContainerLowDark = Color(0xFF1B1C15)
+val surfaceContainerDark = Color(0xFF1F2019)
+val surfaceContainerHighDark = Color(0xFF292B23)
+val surfaceContainerHighestDark = Color(0xFF34352D)
