+++ composeApp/src/desktopMain/kotlin/icu/bluedream/gamesteup/ui/theme/Type.kt
@@ -1,0 +1,5 @@
+package icu.bluedream.gamesteup.ui.theme
+
+import androidx.compose.material3.Typography
+
+val AppTypography = Typography()
